in progress

+++ src/main/java/com/odian/moviesearch/core/model/PagedResponse.java
@@ -5,5 +5,5 @@ package com.odian.moviesearch.core.model;
 import java.util.List;
 
 
-public record PagedResponse<T>(int totalItems, int totalPages, int currentPage, int pageSize, List<T> items) {
+public record PagedResponse<T>(long totalItems, int totalPages, int currentPage, int pageSize, List<T> items) {
 }
